Remove CEF1 specific code since we only focus on CEF3.
